Organize Express routes into a 'routes' directory
