+++ graphql-server/src/routes/auth/check-login.ts
@@ -1,0 +1,6 @@
+const passport = require("passport");
+
+// for navbar to display user email is logged in
+export function checkLogInRouter(req: any, res: any, next: any) {
+  res.status(200).send(req.user.email);
+}
