+++ graphql-server/src/routes/auth/logout.ts
@@ -1,0 +1,10 @@
+const passport = require("passport");
+
+export function logOutRouter(
+  req: any,
+  res: any,
+  next: any,
+  logOutEverywhere: any
+) {
+  logOutEverywhere(req, res, next);
+}
