+++ graphql-server/src/routes/health.ts
@@ -1,0 +1,3 @@
+export function healthCheckRouter(req: any, res: any, next: any) {
+  res.sendStatus(200);
+}
